requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
six==1.14.0
Werkzeug==1.0.0
